Update Admin Main Page Interface to Include Quarter/Year Message
If no quarter/year is set, admins will now see a message on the admin
landing page informing them that a quarter/year needs to be set. Users
will no longer be able to upload a roster if a quarter/year isn't set.
The default assignment number on the add assignment page is now '1' if
no assignments are currently in the DB. Quarters/years are now deleted by
the 'ClearDatabase' handler.